+++ src/main/java/laddergame/controller/LadderController.java
@@ -1,10 +1,9 @@
 package laddergame.controller;
 
-import laddergame.domain.LadderGame;
 import laddergame.domain.LadderHeight;
 import laddergame.domain.Names;
-import laddergame.domain.RandomBooleanGenerator;
 import laddergame.domain.Result;
+import laddergame.service.LadderGame;
 import laddergame.view.InputView;
 import laddergame.view.OutputView;
 
@@ -12,18 +11,19 @@ public class LadderController {
 
     private final InputView inputView;
     private final OutputView outputView;
+    private final LadderGame ladderGame;
 
-    public LadderController(final InputView inputView, final OutputView outputView) {
+    public LadderController(final InputView inputView, final OutputView outputView, final LadderGame ladderGame) {
         this.inputView = inputView;
         this.outputView = outputView;
+        this.ladderGame = ladderGame;
     }
 
     public void run() {
         final Names names = new Names(inputView.readNames());
         final LadderHeight height = new LadderHeight(inputView.readLadderHeight());
 
-        final LadderGame ladderGame = LadderGame.create(names, height, new RandomBooleanGenerator());
-        final Result result = ladderGame.getResult();
+        final Result result = ladderGame.createLadder(names, height);
 
         outputView.printResult(result);
     }
