+++ src/main/java/laddergame/domain/Result.java
@@ -4,5 +4,7 @@ import java.util.List;
 
 public record Result(List<String> names, List<List<Boolean>> ladder) {
 
-
+    public static Result of(final Names names, final Ladder ladder) {
+        return new Result(names.getNames(), ladder.getLinesState());
+    }
 }
